refactor: 오더 페이지에서 state가 최상단에 있으니까 아래 컴포넌트 memoize 및 UI 상관 없는 부분 훅으로 분리
